% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.R
\name{cdc_agree}
\alias{cdc_agree}
\title{Agree to the CDC rules about data usage and confidentiality.}
\usage{
cdc_agree(agree = TRUE)
}
\arguments{
\item{agree}{Logical indicating agreement for TRUE and non-agreement for FALSE}
}
\value{

}
\description{
To use the CDC data you must agree to the terms of use. Do this once per session or you can
add it to individual requests to use programmatically.
}
\examples{
cdc_agree()
cdc_agree(FALSE)
}
\keyword{agreement}
