% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output.R
\name{create_output}
\alias{create_output}
\title{Turns wondr code into a function.}
\usage{
create_output(query_result)
}
\arguments{
\item{query_result}{An object containing the payload from the query}
}
\description{
Turns wondr code into a function.
}
