% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBirths.R
\name{getBirths}
\alias{getBirths}
\title{Request birth data from the CDC Wonder API}
\usage{
getBirths(...)
}
\arguments{
\item{...}{list of words to indicate the group_by variables desired in the query. Words will be interpreted according to the first letter (not case-sensitive):

    A = "Age of Mother Year"

    B = "Birth Weight 100 gram   increments"

    D = "Delivery Method"

    E = "Education"

    G = "Gender"

    H = "Hispanic Origin"

    M = "Month"

    O = "Live Birth Order"

    P = "Birthplace"

    R = "Race"

    S = "Marital Status"

    P = "Month Prenatal Care Began"

    W = "Weekday"}
}
\description{
wrapper to `getData()` which calls the \href{https://wonder.cdc.gov}{CDC Wonder API} and returns a tidy data frame (tibble)
}
\section{Notes}{

`getBirths()` is designed to be easier to use than `getData()` for simple queries. For more options, use `getData()`.
}

\examples{

mydata <- getBirths(age, educcation, wieght, gender)

mydata
}
