% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Request data from the CDC Wonder API}
\usage{
getData(
  db = "D66",
  querylist = NULL,
  add = TRUE,
  save = FALSE,
  fn = "query.xml",
  agree = TRUE
)
}
\arguments{
\item{db}{Indicate the database, either by label, name, or code.  To see the complete list of what's available, use `show_databases()`}

\item{querylist}{The query list supplied must be a list of lists of names and values. By default it is combined with the default query list for that database. To supply a full list and bypass the default query list, set \code{add} to \code{TRUE}. See here for instructions on how to create this list.}

\item{add}{If \code{TRUE} (default) \code{querylist} is combined with the default query list. Set to \code{FALSE} to use \code{querylist} as a standalone list of query parameters.}

\item{save}{If \code{TRUE} the query list will be saved, defaults to \code{FALSE}}

\item{fn}{filename for saved query list, defaults to \code{query.xml}.}

\item{agree}{Must be set to TRUE to agree to the CDC terms (default is TRUE). To see the terms for the particular dataset click on that dataset here: \url{https://wonder.cdc.gov}. \href{https://wonder.cdc.gov/ucd-icd10.html}{sample}}
}
\description{
calls the \href{https://wonder.cdc.gov}{CDC Wonder API} and returns a tidy data frame (tibble)
}
\details{
Queries for mortality and births statistics from the National Vital Statistics System cannot limit or group results by any location field, such as Region, Division, State or County, or Urbanization (urbanization categories map to specific geographic counties).

For example, in the D76 online database for Detailed Mortality 1999-2013, the location fields are D76.V9, D76.V10 and D76.V27, and the urbanization fields are D76.V11 and D76.V19. These 'sub-national" data fields cannot be grouped by or limited via the API, although these fields are available in the web application.

See \url{https://wonder.cdc.gov/wonder/help/wonder-api.html} for more information.
}
\section{References}{

Inspired by this script from the \code{wondr} package, but provides more user-friendly options and better table display: \url{https://github.com/hrbrmstr/wondr/blob/master/README.md}
}

\examples{
\dontrun{
mylist <- list(
  list("And By", "Race"),
  list("And By", "Tobacco Use")
)

mydata <- getData("D66", mylist)
mydata
}

}
