% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_query_list.R
\name{make_query_list}
\alias{make_query_list}
\title{Create the default query request list for a CDC Wonder database call}
\usage{
make_query_list(filename)
}
\arguments{
\item{filename}{the name of the .xml request file that is used as the query default list for that database. The file name must begin with the database code followed by underscore (_), such as D66_sample.xml, and must be located in the data-raw/ directory.}
}
\description{
Query requests must contain many parameters that are generally not of much interest to the end user. Therefore, this package relies on a set of default query lists which the user may alter. This function converts .xml examples from the CDC to R lists. These lists are named by the database code + querydefaults, such as D66querydefaults. Most users will not need to use this function.
}
\examples{
make_query_list("D66_BirthsOnly-req.xml")
make_query_list("D76_Mortality-req.xml")
}
