% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_query.R
\name{send_query}
\alias{send_query}
\title{Send XML query}
\usage{
send_query(db, fn, agree = TRUE)
}
\arguments{
\item{db}{Indicate the database by code}

\item{fn}{\code{.xml} query list exported from CDC Wonder web interface}

\item{agree}{if \code{TRUE} (default) accept CDC Wonder API data use restrictions}
}
\description{
Based on https://github.com/hrbrmstr/wondr/blob/master/R/wondr.r
}
\examples{
\dontrun{
   send_query("D66", "data-raw/D66_Defaults.xml")
   send_query("D76", "testquery.xml")
}

}
