% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_databases.R
\name{show_databases}
\alias{show_databases}
\title{Show available databases}
\usage{
show_databases()
}
\description{
displays available databases (also available here: LINK TO GITHUB)
}
\examples{
show_databases()

}
