% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonderapi-package.R
\docType{package}
\name{wonderapi-package}
\alias{wonderapi}
\alias{wonderapi-package}
\title{wonderapi: User Friendly Access to CDC Wonder API}
\description{
Provides easy access to CDC WONDER data primarily for sociological and
public health topics. Builds on the `wondr` package's general approach
with more specific, human-readable functions to build queries, retrieve
data, and prepare results for reports or further analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/socdataR/wonderapi}
  \item Report bugs at \url{https://github.com/socdataR/wonderapi/issues}
}

}
\author{
\strong{Maintainer}: Joyce Robbins \email{joycerobbins1@gmail.com}

}
\keyword{internal}
