% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_query.R
\name{send_query}
\alias{send_query}
\title{Send XML query}
\usage{
send_query(filename, database_id = NULL)
}
\arguments{
\item{filename}{.xml query list}

\item{database_id}{CDC WONDER database ID, if NULL parsed from filename}
}
\description{
Based on https://github.com/hrbrmstr/wondr/blob/master/R/wondr.r
}
\examples{
\dontrun{
   send_query("data-raw/D66_Defaults.xml")
}
}
