% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonderapi-package.R
\docType{package}
\name{wonderapi-package}
\alias{wonderapi-package}
\alias{wonderapi}
\title{wonderapi}
\description{
It provides easy access to CDC wonder data primarily for sociological topics. It leverages the wondr package's
 general approach to make a more specific and human readable set of functions to build queries, retrieve data, and
 prepare data for either reports or further analysis.
}
\keyword{package}
